# Conflicts:
#	dashboard-spring/src/main/java/com/humintecTest/dashboard/controller/StorageTableController.java
#	dashboard-spring/src/main/java/com/humintecTest/dashboard/controller/StorageUseController.java

+++ dashboard-spring/src/main/java/com/humintecTest/dashboard/response/storageDateResponseFormat.java
@@ -9,7 +9,7 @@ import java.sql.Date;
 @Data
 public class storageDateResponseFormat {
     private Date date;
-    private BigInteger daily_sum;
+    private double daily_sum;
     private String storage_name;
     private BigInteger total;
