+++ dashboard-spring/src/main/java/com/humintecTest/dashboard/vo/ScannerSlideVo.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ScannerSlideVo {
     private String scanner_type;
-    private int slide_sum;
+    private long slide_sum;
 }
